+++ QFAEngine/Engine/Render/RenderPass/RenderPass.cpp
@@ -37,7 +37,7 @@ QFAVKRenderPass::QFAVKRenderPass(VkFormat imageFormat, bool present)
 
     VkSubpassDescription subpass{};
     subpass.pipelineBindPoint = VK_PIPELINE_BIND_POINT_GRAPHICS;
-    subpass.colorAttachmentCount = colorAttachmentRefs.size();
+    subpass.colorAttachmentCount = (uint32_t)colorAttachmentRefs.size();
     subpass.pColorAttachments = colorAttachmentRefs.data();
     subpass.pDepthStencilAttachment = &depthAttachmentRef;
